chore(*): separate dir for exception/viewmodels
