+++ Micro.Auth.Api/Authentication/Exceptions/BadBasicAuthorizationDataException.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace Micro.Auth.Api.Authentication
+namespace Micro.Auth.Api.Authentication.Exceptions
 {
     public class BadBasicAuthorizationDataException : Exception
     {
