+++ Micro.Auth.Api/Authentication/SessionController.cs
@@ -3,9 +3,11 @@ using System.ComponentModel.DataAnnotations;
 using System.Text;
 using System.Threading.Tasks;
 using App.Metrics;
+using Micro.Auth.Api.Authentication.Exceptions;
 using Micro.Auth.Api.Measurements;
 using Micro.Auth.Api.UserData.Extensions;
 using Micro.Auth.Api.Users;
+using Micro.Auth.Api.Users.ViewModels;
 using Micro.Auth.Api.ValidationAttributes;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
