+++ Micro.Auth.Api/RefreshTokens/Exceptions/RefreshTokenNotFoundException.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace Micro.Auth.Api.RefreshTokens
+namespace Micro.Auth.Api.RefreshTokens.Exceptions
 {
     public class RefreshTokenNotFoundException : Exception
     {
